+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- **Dedicated Distribution Module:** Introduced a separate `distributions.py` module to handle the logic for detecting the installed Brave Browser version on different operating systems. This improves code organization and maintainability.
+
 ## [0.1.8] - 2025-05-10
 ### Fixed
 - **Fixed for config path:** Corrected to check config.ini file if exists instead the path of the config.ini.